auth/tests/functional/requirements.txt
pydantic==1.9.0
pytest==7.1.1
pytest-asyncio==0.18.3
aioredis==2.0.1
aiohttp==3.8.1
aiopg==1.3.3
psycopg2-binary==2.9.3
requests==2.27.1
redis==4.2.2
pydantic==1.9.0
Flask-Migrate==3.1.0
grpcio==1.46.0
grpcio-tools==1.46.0